Move parsing code into parser
This involves a bit more indirection in the macro.

The function `index-variables` previously associated with each variable
in each clause a function that would get that variable out of the
clause.

This to me goes slightly beyond the work of a parser. The parser should
just say what position the variable is in. It's then the work of the
compiler to do the right thing. It's just a simple layer of indirection
from a keyword denoting the position to a function for getting something
in that position, but it seemed right.

It might be a bit pedantic, but I thought it was worth it for the
practice at macro-fu.