WIP on handling predicates
It pretty much works - except when it totally doesn't.

I implemented a pretty naive translation of the predicate
(like `[(< ?danny-born ?actor-born)]` to a function that substitutes
any variables for their values out of the zset item. It seems to work!
But there is a type problem: ?danny-born is coming through as a
java.util.Date, and Clojure's < operator isn't defined over dates.

Seems like = and not= should be ok.