Using new support in group-by-xf for tagging aggregates with variables
Until now, group-by-xf did not support multiple aggregates of the same
type on the same query. Now, you pass in the variable being aggregated
for each one, and it comes out in the query result.

With this, I'd still like to clean up the API for aggregate queries
(such as the need to call get-aggregate-result instead of merely
get-result), but we can merge this, and I think we can say aggregates
are supported!