POC for rendering aggregate results as vector preserving query order
Aggregate query results are maintained in an indexed zset.
Each "row" of the result has an entry in the map, where the key is the
group-by key, and the value is a zset.

Users probably expect the query result to come in the sequence that the
variables appeared in the :find in the query. Zsets don't have an
ordering, so the answers don't come out in any particular order.
Currently though we are taggging them with the corresponding variable.

In other words, if your query begins :find ?a (count ?b) (sum ?c), the
results are of the form #{ [?b num1] [?c num2 }, where [?a] was the key
in the map. Datascript returns results in the form [?a num1 num2].

This change gets most of the way to matching datascript's output, by:
- putting the query itself on the metadata of the compiled xf
- putting that query in the zsxf query map that's generated
- when get-aggregate-result is called, mapping over the saved aggregate
  query parameters to get the corresponding values out

So, once again, instead of results of the form {[?a] #{[?c num2] [?b num1]},
we now have {[?a] [num1 num2]}.

This is just a POC, because I don't think the methods used here (and the
branch in logic for get-aggregate-xf) are what we'll want long term.