Improve compiler exception when passed invalid or unsupported query
It now returns code which throws, instead of the literal exception.

Macroexpanding the offending query makes it easy to see the spec
problem, and executing the macroexpanded code throws the right error
(fixed a bug with quoting the ex-data).

There's possibly more to be done to provide a better u/x for
understanding spec failures.